requirements.txt
Flask==2.3.3
numpy==1.24.3
opencv-python-headless==4.8.1.78
tensorflow-cpu==2.16.1
keras==3.0.5
matplotlib==3.7.2
Pillow==10.0.0
gunicorn==21.2.0